Build ninja using Kokoro
This script builds ninja on Linux, Windows, and macOS. It produces a
zip file containing:

 * ninja[.exe]
 * ninja/COPYING (the Apache2 license)

On Windows, this script builds the binary using the "Visual Studio 15
2017 Win64" generator. ninja's CMakeLists.txt file sets
CMAKE_MSVC_RUNTIME_LIBRARY to MultiThreaded (as opposed to
MultiThreadedDLL), so ninja.exe depends only on kernel32.dll.

On macOS, this script targets 10.9 and up, and generates a universal
arm64+x86_64 binary.

On Linux, this script uses g++ 4.8 from CentOS 7 and links with
-static-libstdc++.

This script runs the ninja tests (ninja_test[.exe]), which complete in
just a few seconds.

Bug: https://github.com/android/ndk/issues/1548
Test: kokoro/kokoro_build.cmd
Test: kokoro/kokoro_build.sh (on Linux and macOS)
Change-Id: If4f7055d8791a5b8687fac14356f59350ae894a4