Add 'sandbox_disabled' variable to opt out of sandboxing
Currently, there is no mechanism for an individual build rule to declare
itself as incompatible and opt out of the global action sandboxing. This
forces an all-or-nothing approach.

This change introduces a new rule variable, sandbox_disabled, to
allow for per-action control over sandboxing.

If a build statement in a .ninja file contains the line
  sandbox_disabled = true
(or any non-empty value), Ninja will now skip wrapping that specific
action in the nsjail sandbox, running it directly instead.

Test: manual
Bug: 454133382
Change-Id: Idb49cfdbf5b93a2cf298d257cd6e62893e7ecf33