Disable nsjail for rules with outputs outside the build dir
Nsjail sandboxing assumes all declared outputs are real files that will be created inside the build directory (e.g., `out/`). This assumption fails for some examples, including:
- phony-like rules, which declare outputs that are not file paths within the build directory.
- dist rules in buildbots, where DIST_DIR is outside of `out/`.

This change fixes the issue by detecting this specific condition.
Before a command is run, Ninja checks if any of its declared outputs have a path that does not start with the configured build directory. If such an output is found, the rule is executed outside of sandbox.

Test: copy ninja to main, SOONG_ACTION_SANDBOXING=nsjail m nothing
Bug: 456291105
Change-Id: Id09ce427624b235eae153a77d4cb2f35d29cdee9