Recursively resolve phony target inputs in sandboxed builds
Sandboxed builds would crash with a fatal "Failed to lstat" error if a
build rule's inputs included a phony target.

This was caused by the sandboxing preparation logic, which incorrectly
assumed that every input node corresponded to a physical file on disk.
It would attempt to call `lstat` on the phony target's path, which
rightfully fails as the file does not exist.

This change fixes the issue by replacing the simple input loop with a
traversal. The new logic recursively expands any phony targets until
it resolves them to non-phony file dependencies.

Test: copy to main, SOONG_ACTION_SANDBOXING=nsjail m
Bug: 456856891
Change-Id: I9b45edd5ad7e761353323b4cb25ebd756537235f