Add /dev mount and enable proc mounting in nsjail
To support build tools that require access to device files and process
information, this change adds a bind mount for /dev and enables the
mounting of /proc within the nsjail sandbox.

This change also fixes several copy-paste errors where the `is_dir`
attribute was being set on the wrong mount object (`binMount`) for the
lib64, usr, tmp, and src mounts.

Test: OUT_DIR=out prebuilts/build-tools/build-prebuilts.sh
Bug: 462186888
Change-Id: Icb12241baf04ccb5983654b00e8e55bc3ea1331e